backend/requirements.txt
# Django 后端（当前兼容 Python 3.11）
Django>=5.2,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.6
python-dotenv>=1.0
psycopg2-binary>=2.9

# 检测与任务管线
opencv-python-headless>=4.8
ultralytics>=8.0

# LLM 调用（ai_core/llm_agent.py）
openai>=1.0

# 可选：仅当你要跑 backend/main.py 里的 FastAPI 示例时才需要
# fastapi
# uvicorn[standard]
